+++ android/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Changed
+* Number field now can include alphabetic characters
+
 ## [1.0.0]
 ### Changed
 * Baseball Card Tracker has a Material Design face lift. All the features you use are still